Moved TryGetDateTimeInput extension into the FluidValueExtensions class along with the private static values it depends on.
